Move TimProduct class into its own file.
